% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{fizz_buzz}
\alias{fizz_buzz}
\title{\title{Fizz Buzz}}
\usage{
fizz_buzz(n)
}
\description{
\description{Any number evenly divisible by 3 is substituted by “Fizz” and any number
evenly divisible by 5 is substituted by “Buzz,” and if it is divisible by both it is substituted by “Fizz Buzz.”}
}
