% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{A time series of daily maximum temperatures in Flagstaff, AZ.}
\format{
a data frame with 22967 observations
\describe{
  \item{Date}{The date of observation as a POSIX date format.}
  \item{Precip}{The amount of precipitation, in inches.}
  \item{SnFall}{Snowfall, in inches.}
  \item{HTemp}{Daily maximum temperature in degrees Farhenheit.}
  \item{LTemp}{Daily minimum temperature in degrees Farhenheit.}
}
}
\source{
\url{https://www.ncdc.noaa.gov/cdo-web/search}
}
\usage{
Flagstaff_Weather
}
\description{
A time series of daily maximum temperatures in Flagstaff, AZ.
}
\keyword{datasets}
